data type add for numpy

+++ deepmd/tf/nvnmd/descriptor/se_atten.py
@@ -61,9 +61,9 @@ def check_switch_range(davg, dstd) -> None:
     #  'init_from_model', 'restart', 'init_from_frz_model', 'finetune'
     if (davg is not None) or (dstd is not None):
         if davg is None:
-            davg = np.zeros([ntype, ndescrpt])
+            davg = np.zeros([ntype, ndescrpt], dtype=np.float64)
         if dstd is None:
-            dstd = np.ones([ntype, ndescrpt])
+            dstd = np.ones([ntype, ndescrpt], dtype=np.float64)
         nvnmd_cfg.get_s_range(davg, dstd)
